+++ src/Elastica.php
@@ -11,8 +11,9 @@
 namespace fork\elastica;
 
 use Craft;
+use craft\base\Element;
+use craft\base\Model;
 use craft\base\Plugin;
-use craft\elements\Entry;
 use craft\events\ModelEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterUserPermissionsEvent;
@@ -22,7 +23,11 @@ use fork\elastica\models\Settings;
 use fork\elastica\services\Indexer;
 use fork\elastica\services\Utility as UtilityService;
 use fork\elastica\utilities\Utility as CpUtility;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Error\SyntaxError;
 use yii\base\Event;
+use yii\base\Exception;
 
 /**
  * Craft plugins are very much like little applications in and of themselves. We’ve made
@@ -38,7 +43,7 @@ use yii\base\Event;
  * @package   Elastica
  * @since     1.0.0
  *
- * @property  \fork\elastica\services\Utility $utility
+ * @property  UtilityService $utility
  * @property  Indexer $indexer
  * @property  Settings $settings
  * @method    Settings getSettings()
@@ -65,7 +70,7 @@ class Elastica extends Plugin
      *
      * @var string
      */
-    public $schemaVersion = '1.0.0';
+    public string $schemaVersion = '2.0.0';
 
     // Public Methods
     // =========================================================================
@@ -101,13 +106,13 @@ class Elastica extends Plugin
         }
 
         // elasticsearch index actions
-        Event::on(Entry::class, Entry::EVENT_AFTER_SAVE, function (ModelEvent $event) {
+        Event::on(Element::class, Element::EVENT_AFTER_SAVE, function (ModelEvent $event) {
             $this->indexer->handleAfterSaveEvent($event);
         });
-        Event::on(Entry::class, Entry::EVENT_AFTER_RESTORE, function (Event $event) {
+        Event::on(Element::class, Element::EVENT_AFTER_RESTORE, function (Event $event) {
             $this->indexer->handleAfterRestoreEvent($event);
         });
-        Event::on(Entry::class, Entry::EVENT_AFTER_DELETE, function (Event $event) {
+        Event::on(Element::class, Element::EVENT_AFTER_DELETE, function (Event $event) {
             $this->indexer->handleAfterDeleteEvent($event);
         });
 
@@ -115,40 +120,19 @@ class Elastica extends Plugin
             UserPermissions::class,
             UserPermissions::EVENT_REGISTER_PERMISSIONS,
             function(RegisterUserPermissionsEvent $event) {
-                $event->permissions['Elastica'] = [
-                    'elasticaIndexTemplates' => [
-                        'label' => 'Set index templates in elasticsearch',
-                    ],
+                $event->permissions[] = [
+                    'heading' => 'Elastica',
+                    'permissions' => [
+                        'elasticaIndexTemplates' => [
+                            'label' => 'Set index templates in elasticsearch',
+                        ],
+                        'elasticaSearchTemplates' => [
+                            'label' => 'Add/update search templates in elasticsearch',
+                        ],
+                    ]
                 ];
             }
         );
-
-        /**
-         * Logging in Craft involves using one of the following methods:
-         *
-         * Craft::trace(): record a message to trace how a piece of code runs. This is mainly for development use.
-         * Craft::info(): record a message that conveys some useful information.
-         * Craft::warning(): record a warning message that indicates something unexpected has happened.
-         * Craft::error(): record a fatal error that should be investigated as soon as possible.
-         *
-         * Unless `devMode` is on, only Craft::warning() & Craft::error() will log to `craft/storage/logs/web.log`
-         *
-         * It's recommended that you pass in the magic constant `__METHOD__` as the second parameter, which sets
-         * the category to the method (prefixed with the fully qualified class name) where the constant appears.
-         *
-         * To enable the Yii debug toolbar, go to your user account in the AdminCP and check the
-         * [] Show the debug toolbar on the front end & [] Show the debug toolbar on the Control Panel
-         *
-         * http://www.yiiframework.com/doc-2.0/guide-runtime-logging.html
-         */
-        //Craft::info(
-        //    Craft::t(
-        //        'elasticsearch',
-        //        '{name} plugin loaded',
-        //        ['name' => $this->name]
-        //    ),
-        //    __METHOD__
-        //);
     }
 
     // Protected Methods
@@ -157,9 +141,9 @@ class Elastica extends Plugin
     /**
      * Creates and returns the model used to store the plugin’s settings.
      *
-     * @return \craft\base\Model|null
+     * @return Model|null
      */
-    protected function createSettingsModel()
+    protected function createSettingsModel(): ?Model
     {
         return new Settings();
     }
@@ -170,16 +154,17 @@ class Elastica extends Plugin
      *
      * @return string The rendered settings HTML
      *
-     * @throws \Twig\Error\LoaderError
-     * @throws \Twig\Error\RuntimeError
-     * @throws \Twig\Error\SyntaxError
-     * @throws \yii\base\Exception
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     * @throws Exception
      */
     protected function settingsHtml(): string
     {
         return Craft::$app->view->renderTemplate(
             'elastica/settings',
             [
+                'connectionStatus' => $this->indexer->getConnectionStatus(),
                 'settings' => $this->getSettings(),
             ]
         );