+++ src/models/Settings.php
@@ -56,6 +56,13 @@ class Settings extends Model
      */
     public $indexTemplate;
 
+    /**
+     * Search templates for elasticsearch
+     *
+     * @var string
+     */
+    public $searchTemplates;
+
     // Public Methods
     // =========================================================================
 
@@ -69,7 +76,7 @@ class Settings extends Model
      *
      * @return array
      */
-    public function rules()
+    public function rules(): array
     {
         return [
             ['hosts', 'required'],
@@ -82,7 +89,23 @@ class Settings extends Model
                     $this->addError($attribute, $exception->getMessage());
                 }
             }],
+            ['searchTemplates', function ($attribute, $params, $validator) {
+                try {
+                    if ($this->$attribute) {
+                        foreach ($this->$attribute as $row) {
+                            if (empty($row[0])) {
+                                throw new \Exception("Handle must not be empty");
+                            }
+                            Json::decode($row[1]);
+                            if ( ! empty($row[2])) {
+                                Json::decode($row[2]);
+                            }
+                        }
+                    }
+                } catch (\Exception $exception) {
+                    $this->addError($attribute, $exception->getMessage());
+                }
+            }],
         ];
     }
 }
-