+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0 - 2024-10-02
+### Added
+- Added support for Craft 4
+- Added optional indexing of categories and assets
+- Added a connection status indicator to the utility
+- Added an option to define search templates
+
+### Changed
+- Include site handle in index name
+
 ## 1.0.2.2 - 2022-02-04
 ### Fixed
 - Fix indexing after restoring trashed entries